refactor: move windows path logic out of `spack.util.path` and into `llnl.util.path`
