+++ lib/spack/llnl/util/filesystem.py
@@ -20,10 +20,11 @@ from sys import platform as _platform
 
 from llnl.util import tty
 from llnl.util.lang import dedupe, memoized
+from llnl.util.path import path_to_os_path, system_path_filter
 from llnl.util.symlink import islink, symlink
 
 from spack.util.executable import CommandNotFoundError, Executable, which
-from spack.util.path import path_to_os_path, system_path_filter
+
 
 is_windows = _platform == "win32"
