+++ lib/spack/spack/compiler.py
@@ -24,7 +24,7 @@ import spack.util.executable
 import spack.util.module_cmd
 import spack.version
 from spack.util.environment import filter_system_paths
-from spack.util.path import system_path_filter
+from llnl.util.path import system_path_filter
 
 __all__ = ["Compiler"]
