+++ lib/spack/spack/spec.py
@@ -62,6 +62,7 @@ import ruamel.yaml as yaml
 import llnl.util.string
 import llnl.util.filesystem as fs
 import llnl.util.lang as lang
+import llnl.util.path as pth
 import llnl.util.tty as tty
 import llnl.util.tty.color as clr
 
@@ -83,7 +84,6 @@ import spack.util.crypto
 import spack.util.executable
 import spack.util.hash
 import spack.util.module_cmd as md
-import spack.util.path as pth
 import spack.util.prefix
 import spack.util.spack_json as sjson
 import spack.util.spack_yaml as syaml
